Create Chineto Sonia Chizitere 2024714027
This Python program implements grading system using "if and elif" statements. It takes a student's score as input and print the corresponding grade based on the following scale:

70 - 100 : Grade A
60 - 69 : Grade B
50 - 59 : Grade C
45 - 44 : Grade D
20 - 39 : Grade F

For each valid score, the program also prints "Yes" to confirm the grading was successful. If the range score is outside the valid range, it's prints "invalid score" and "No".

Created by: Chineto Sonia Chizitere 
Reg No : 2024714027